requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Deprecated==1.2.14
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.10
Markdown==3.4.4
multidict==6.0.4
pycparser==2.21
PyGithub==1.59.0
PyJWT==2.8.0
PyNaCl==1.5.0
pytz==2023.3
requests==2.31.0
urllib3==2.0.4
wrapt==1.15.0
yarl==1.9.2
